refactor(reports): extract detailed report to reports_detailed.go
